Collision is now applied to indexes i and j (see code)
This will stop weird bugs where player will blink and enemy still chases.


On an extreme side note, notice that I put i-- and j-- when GameObjects.RemoveAt is called. This is because when an item is removed AND i is incremented, you skip the following element (for one update call)